# bpp/backend/docs/swagger.yaml
definitions:
  business.AddBusinessRequest:
    properties:
      description:
        type: string
      email:
        type: string
      gstIndexNumber:
        type: string
      industry:
        $ref: '#/definitions/business.Industry'
      location:
        $ref: '#/definitions/business.Location'
      name:
        type: string
      phone:
        type: string
      pictureUrls:
        items:
          type: string
        type: array
    type: object
  business.Coordinates:
    properties:
      coordinates:
        description: longitude, latitude
        items:
          type: number
        type: array
      type:
        description: GeoJSON type, it shall be equal to 'Point'
        type: string
    type: object
  business.Industry:
    enum:
    - RetailAndEcommerce
    - FoodAndBeverages
    - HealthAndWellness
    - EducationAndTraining
    - ProfessionalServices
    - Manufacturing
    - HospitalityAndTourism
    - ArtsAndEntertainment
    - TechnologyAndSoftware
    - ConstructionAndRealEstate
    - TransportationAndLogistics
    - AgricultureAndFarming
    - FinanceAndInsurance
    - EnergyAndUtilities
    - NonProfitAndSocialEnterprise
    - MediaAndPublishing
    - Automotive
    - FashionAndLifestyle
    - SportsAndRecreation
    - Other
    type: string
    x-enum-varnames:
    - IndustryRetailAndEcommerce
    - IndustryFoodAndBeverages
    - IndustryHealthAndWellness
    - IndustryEducationAndTraining
    - IndustryProfessionalServices
    - IndustryManufacturing
    - IndustryHospitalityAndTourism
    - IndustryArtsAndEntertainment
    - IndustryTechnologyAndSoftware
    - IndustryConstructionAndRealEstate
    - IndustryTransportationAndLogistics
    - IndustryAgricultureAndFarming
    - IndustryFinanceAndInsurance
    - IndustryEnergyAndUtilities
    - IndustryNonProfitAndSocialEnterprise
    - IndustryMediaAndPublishing
    - IndustryAutomotive
    - IndustryFashionAndLifestyle
    - IndustrySportsAndRecreation
    - IndustryOther
  business.ListJobsResponse:
    properties:
      applicationIds:
        items:
          type: string
        type: array
      description:
        type: string
      eligibility:
        $ref: '#/definitions/job.Eligibility'
      id:
        type: string
      location:
        $ref: '#/definitions/job.Location'
      name:
        type: string
      salaryRange:
        $ref: '#/definitions/job.SalaryRange'
      type:
        $ref: '#/definitions/job.JobType'
      vacancies:
        type: integer
      workDays:
        $ref: '#/definitions/job.WorkDays'
      workHours:
        $ref: '#/definitions/job.WorkHours'
    type: object
  business.Location:
    properties:
      address:
        type: string
      city:
        description: 'STD code, for example: ''std:080'''
        type: string
      coordinates:
        $ref: '#/definitions/business.Coordinates'
      postalCode:
        description: 'Postal code, for example: ''560102'''
        type: string
      state:
        description: 'State code, for example: ''IN-KA'''
        type: string
      street:
        type: string
    type: object
...
  job.AcademicQualification:
    enum:
    - None
    - Class-X
    - Class-XII
    - Diploma
    - Graduate
    - Post-Graduate
    type: string
    x-enum-varnames:
    - AcademicQualificationNone
    - AcademicQualificationClassX
    - AcademicQualificationClassXII
    - AcademicQualificationDiploma
    - AcademicQualificationGraduate
    - AcademicQualificationPostGraduate
  job.Coordinates:
    properties:
      coordinates:
        description: longitude, latitude
        items:
          type: number
        type: array
      type:
        description: GeoJSON type, it shall be equal to 'Point'
        type: string
    type: object
...
  job.Document:
    enum:
    - aadhar_card
    - pan_card
    - driving_license
    - class_x_certificate
    - class_xii_certificate
    - diploma_certificate
    - graduation_certificate
    - post_grad_certificate
    - passport
    - other
    type: string
    x-enum-varnames:
    - DocumentAadharCard
    - DocumentPanCard
    - DocumentDrivingLic
    - DocumentClassXCert
    - DocumentClassXIICertificate
    - DocumentDiplomaCertificate
    - DocumentGraduationCertificate
    - DocumentPostGradCertificate
    - DocumentPassport
    - DocumentOther
  job.Eligibility:
    properties:
      academicQualification:
        $ref: '#/definitions/job.AcademicQualification'
      documentsRequired:
        items:
          $ref: '#/definitions/job.Document'
        type: array
      gender:
        $ref: '#/definitions/job.Gender'
      yearsOfExperience:
        type: integer
    type: object
  job.Gender:
    enum:
    - any
    - male
    - female
    type: string
    x-enum-varnames:
    - GenderAny
    - GenderMale
    - GenderFemale
...
  job.JobType:
    enum:
    - full-time
    - part-time
    - contract
    - internship
    type: string
    x-enum-varnames:
    - JobTypeFullTime
    - JobTypePartTime
    - JobTypeContract
    - JobTypeInternship
  job.Location:
    properties:
      address:
        type: string
      city:
        description: 'STD code, for example: ''std:080'''
        type: string
      coordinates:
        $ref: '#/definitions/job.Coordinates'
      postalCode:
        description: 'Postal code, for example: ''560102'''
        type: string
      state:
        description: 'State code, for example: ''IN-KA'''
        type: string
      street:
        type: string
    type: object
  job.SalaryRange:
    properties:
      max:
        type: integer
      min:
        type: integer
    type: object
  job.WorkDays:
    properties:
      end:
        type: integer
      start:
        type: integer
    type: object
  job.WorkHours:
    properties:
      end:
        type: string
      start:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /business/{id}/jobs:
    get:
      consumes:
      - application/json
      description: List jobs for a business
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/business.ListJobsResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List Jobs
      tags:
      - Business
  /business/add:
    post:
      consumes:
      - application/json
      description: Add a business
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/business.AddBusinessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Add business
      tags:
      - Business
...
swagger: "2.0"